# openapi.yaml
openapi: 3.0.3
info:
  title: Flashcards API
  description: |
    AI-engineered flashcard application for language learning with intelligent translation features.
    
    This API provides a stateful workflow for creating flashcards with optional AI-powered translations.
    Generate translations, review them, and then save your flashcards with full control over the content.
    
    **Key Features:**
    - 🤖 **AI Translation Generation** - Generate translations on-demand (cached for 1 hour)
    - 📝 **Manual Card Creation** - Full control over both sides of the card
    - 🎯 **Translation Tracking** - Know which fields were AI-generated
    - 🌍 **Multi-language Support** - Any ISO 639-1 language pair
    - 📚 **Study Mode** - Random flashcard selection with optional AI hints
    - 🔄 **Full CRUD** - Create, read, update, and delete flashcards
    
    **Workflow:**
    1. User types content in one language
    2. Call `/flashcards/translate` to generate translation (cached)
    3. Review/edit the suggested translation
    4. Call `/flashcards` to save the card with AI tracking flags
    
    **GitHub Repository:** [https://github.com/akolybelnikov/flashcards](https://github.com/akolybelnikov/flashcards)
  version: 2.0.0
  contact:
    name: API Support
    url: https://github.com/akolybelnikov/flashcards/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://flashcards-hqwc.onrender.com
    description: Production server
  - url: http://localhost:10000
    description: Local development server

tags:
  - name: Flashcards
    description: CRUD operations for managing flashcards
  - name: Translation
    description: AI-powered translation generation with caching
  - name: Health
    description: Health check endpoints

paths:
  /health:
    get:
      summary: Health check
      description: Check if the API server is running and healthy
      tags:
        - Health
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
              example:
                status: "healthy"

  /flashcards:
    get:
      summary: Get all flashcards
      description: Retrieve a list of all flashcards in the system
      tags:
        - Flashcards
      responses:
        '200':
          description: List of flashcards retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Flashcard'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new flashcard with optional AI translation
      description: |
        Create a new flashcard by providing either question OR answer, or both.
        
        **Three Creation Modes:**
        
        1. **Both Fields Provided** - No AI translation needed
           - Provide both `question` and `answer`
           - `from_lang` and `to_lang` are optional
           - Card is created directly with both fields
        
        2. **Only Question Provided** - AI translates to answer
           - Provide `question`, leave `answer` empty
           - Must provide `from_lang` and `to_lang`
           - AI generates translation for the answer
           - `ai_translated_answer` will be set to `true`
        
        3. **Only Answer Provided** - AI translates to question
           - Provide `answer`, leave `question` empty
           - Must provide `from_lang` and `to_lang`
           - AI generates translation for the question
           - `ai_translated_question` will be set to `true`
        
        **Caching:**
        - Translations are cached for 1 hour
        - Same content + language pair = instant response
        - Concurrent requests are deduplicated
        
        **Language Codes:**
        Use ISO 639-1 codes (e.g., "en", "el", "fr").
        Required when translation is needed, optional otherwise.
      tags:
        - Flashcards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFlashcardRequest'
            examples:
              bothFieldsManual:
                summary: Both fields provided (no AI)
                value:
                  question: "hello"
                  answer: "γεια σας"
                  from_lang: "en"
                  to_lang: "el"
              onlyQuestion:
                summary: Only question (AI translates to answer)
                value:
                  question: "hello"
                  answer: ""
                  from_lang: "en"
                  to_lang: "el"
              onlyAnswer:
                summary: Only answer (AI translates to question)
                value:
                  question: ""
                  answer: "γεια σας"
                  from_lang: "el"
                  to_lang: "en"
      responses:
        '201':
          description: Flashcard created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
              examples:
                manual:
                  summary: Manual creation (no AI)
                  value:
                    id: 1
                    question: "hello"
                    answer: "γεια σας"
                    question_lang: "en"
                    answer_lang: "el"
                    ai_translated_question: false
                    ai_translated_answer: false
                    created_at: "2025-11-02T10:00:00Z"
                    updated_at: "2025-11-02T10:00:00Z"
                aiAnswer:
                  summary: AI-generated answer
                  value:
                    id: 2
                    question: "hello"
                    answer: "γεια σας"
                    question_lang: "en"
                    answer_lang: "el"
                    ai_translated_question: false
                    ai_translated_answer: true
                    created_at: "2025-11-02T10:00:00Z"
                    updated_at: "2025-11-02T10:00:00Z"
                aiQuestion:
                  summary: AI-generated question
                  value:
                    id: 3
                    question: "hello"
                    answer: "γεια σας"
                    question_lang: "en"
                    answer_lang: "el"
                    ai_translated_question: true
                    ai_translated_answer: false
                    created_at: "2025-11-02T10:00:00Z"
                    updated_at: "2025-11-02T10:00:00Z"
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                bothEmpty:
                  summary: Both fields empty
                  value:
                    error: "At least one of question or answer must be provided"
                missingLanguages:
                  summary: Missing language codes when translation needed
                  value:
                    error: "from_lang and to_lang are required when translation is needed"
                noApiKey:
                  summary: OpenAI not configured
                  value:
                    error: "AI translation not available: API key not configured"
                translationFailed:
                  summary: Translation generation failed
                  value:
                    error: "failed to generate answer translation: API error"
                invalidJson:
                  summary: Invalid JSON
                  value:
                    error: "Invalid JSON payload"

  /flashcards/{id}:
    get:
      summary: Get a flashcard by ID
      description: Retrieve a specific flashcard by its ID
      tags:
        - Flashcards
      parameters:
        - name: id
          in: path
          required: true
          description: Flashcard ID
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        '200':
          description: Flashcard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
        '400':
          description: Invalid flashcard ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "flashcard with id not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a flashcard
      description: Update the question and/or answer of an existing flashcard
      tags:
        - Flashcards
      parameters:
        - name: id
          in: path
          required: true
          description: Flashcard ID
          schema:
            type: integer
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFlashcardRequest'
            examples:
              updateBoth:
                summary: Update both fields
                value:
                  question: "goodbye"
                  answer: "αντίο"
              updateQuestion:
                summary: Update only question
                value:
                  question: "goodbye"
              updateAnswer:
                summary: Update only answer
                value:
                  answer: "αντίο"
      responses:
        '200':
          description: Flashcard updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
        '400':
          description: Bad request (invalid ID or payload)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a flashcard
      description: Delete a flashcard by its ID
      tags:
        - Flashcards
      parameters:
        - name: id
          in: path
          required: true
          description: Flashcard ID
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        '204':
          description: Flashcard deleted successfully (no content)
        '400':
          description: Invalid flashcard ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /flashcards/translate:
    post:
      summary: Generate AI translation
      description: |
        Generate an AI-powered translation for flashcard content.
        
        **Caching:**
        - First request: Calls OpenAI API (slower, ~1-3 seconds)
        - Subsequent requests: Returns cached result (fast, <10ms)
        - Cache TTL: 1 hour (configurable)
        - Cache key: Hash of (content + from_lang + to_lang)
        
        **Workflow:**
        1. User types "hello" in English field
        2. Click "Generate Translation" → Call this endpoint
        3. Backend returns "γεια σας" with `cached: false`
        4. User can accept, edit, or discard
        5. Clicking "Generate" again returns same translation with `cached: true`
        
        **Language Codes:**
        Use ISO 639-1 codes (2 lowercase letters). Examples:
        - `en` = English
        - `el` = Greek (Ελληνικά)
        - `fr` = French
        - `de` = German
        - `es` = Spanish
        
        See [ISO 639-1 codes](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) for full list.
        
        **Note:** Requires `OPENAI_API_KEY` to be configured on the server.
      tags:
        - Translation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTranslationRequest'
            examples:
              englishToGreek:
                summary: English to Greek
                value:
                  content: "hello"
                  from_lang: "en"
                  to_lang: "el"
              greekToEnglish:
                summary: Greek to English
                value:
                  content: "γεια σας"
                  from_lang: "el"
                  to_lang: "en"
              frenchToEnglish:
                summary: French to English
                value:
                  content: "bonjour"
                  from_lang: "fr"
                  to_lang: "en"
      responses:
        '200':
          description: Translation generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTranslationResponse'
              examples:
                cacheMiss:
                  summary: First request (cache miss)
                  value:
                    translation: "γεια σας"
                    cached: false
                    cache_key: "a1b2c3d4e5f6..."
                cacheHit:
                  summary: Subsequent request (cache hit)
                  value:
                    translation: "γεια σας"
                    cached: true
                    cache_key: "a1b2c3d4e5f6..."
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                emptyContent:
                  summary: Missing content
                  value:
                    error: "Content cannot be empty"
                missingLanguages:
                  summary: Missing language codes
                  value:
                    error: "Both from_lang and to_lang must be provided"
                invalidLanguageCode:
                  summary: Invalid language format
                  value:
                    error: "invalid source language: language code must be 2 characters"
                aiNotAvailable:
                  summary: OpenAI not configured
                  value:
                    error: "AI translation not available: API key not configured"
        '500':
          description: Translation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "translation failed: API error"

  /flashcards/random:
    get:
      summary: Get a random flashcard
      description: |
        Retrieve a random flashcard for study purposes, optionally with an AI-generated hint.
        
        The `lang` query parameter specifies the language for the AI hint. If omitted, the hint
        may be generated in a default language or not generated at all.
      tags:
        - Flashcards
      parameters:
        - name: lang
          in: query
          required: false
          description: Language code for the AI hint (e.g., 'el' for Greek, 'en' for English)
          schema:
            type: string
            enum: [en, el]
            example: el
      responses:
        '200':
          description: Random flashcard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RandomFlashcardResponse'
              examples:
                withHint:
                  summary: With AI hint
                  value:
                    flashcard:
                      id: 5
                      question: "hello"
                      answer: "γεια σας"
                      question_lang: "en"
                      answer_lang: "el"
                      ai_translated_question: false
                      ai_translated_answer: true
                      created_at: "2025-11-02T10:00:00Z"
                      updated_at: "2025-11-02T10:00:00Z"
                    ai_hint: "A common Greek greeting used in formal situations"
                withoutHint:
                  summary: Without AI hint
                  value:
                    flashcard:
                      id: 5
                      question: "hello"
                      answer: "γεια σας"
                      question_lang: "en"
                      answer_lang: "el"
                      ai_translated_question: false
                      ai_translated_answer: false
                      created_at: "2025-11-02T10:00:00Z"
                      updated_at: "2025-11-02T10:00:00Z"
                    ai_hint: null
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Flashcard:
      type: object
      required:
        - id
        - question
        - answer
        - ai_translated_question
        - ai_translated_answer
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the flashcard
          example: 1
        question:
          type: string
          description: The question/front side of the flashcard
          example: "hello"
        answer:
          type: string
          description: The answer/back side of the flashcard
          example: "γεια σας"
        question_lang:
          type: string
          nullable: true
          description: ISO 639-1 language code for the question (e.g., "en", "el")
          pattern: '^[a-z]{2}$'
          example: "en"
        answer_lang:
          type: string
          nullable: true
          description: ISO 639-1 language code for the answer (e.g., "en", "el")
          pattern: '^[a-z]{2}$'
          example: "el"
        ai_translated_question:
          type: boolean
          description: Whether the question was AI-generated
          example: false
        ai_translated_answer:
          type: boolean
          description: Whether the answer was AI-generated
          example: true
        created_at:
          type: string
          format: date-time
          description: Timestamp when the flashcard was created
          example: "2025-11-02T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the flashcard was last updated
          example: "2025-11-02T10:00:00Z"

    CreateFlashcardRequest:
      type: object
      properties:
        question:
          type: string
          description: |
            The question/front side of the flashcard.
            Required if `answer` is empty (AI will translate it).
            Optional if `answer` is provided.
          example: "hello"
        answer:
          type: string
          description: |
            The answer/back side of the flashcard.
            Required if `question` is empty (AI will translate it).
            Optional if `question` is provided.
          example: "γεια σας"
        from_lang:
          type: string
          description: |
            ISO 639-1 language code of the provided content.
            Required when translation is needed (one field is empty).
            Optional when both fields are provided.
          pattern: '^[a-z]{2}$'
          example: "en"
        to_lang:
          type: string
          description: |
            ISO 639-1 language code to translate to.
            Required when translation is needed (one field is empty).
            Optional when both fields are provided.
          pattern: '^[a-z]{2}$'
          example: "el"
      example:
        question: "hello"
        answer: ""
        from_lang: "en"
        to_lang: "el"

    GenerateTranslationRequest:
      type: object
      required:
        - content
        - from_lang
        - to_lang
      properties:
        content:
          type: string
          description: The text content to translate
          minLength: 1
          example: "hello"
        from_lang:
          type: string
          description: Source language code (ISO 639-1, 2 lowercase letters)
          pattern: '^[a-z]{2}$'
          example: "en"
        to_lang:
          type: string
          description: Target language code (ISO 639-1, 2 lowercase letters)
          pattern: '^[a-z]{2}$'
          example: "el"
      example:
        content: "hello"
        from_lang: "en"
        to_lang: "el"

    GenerateTranslationResponse:
      type: object
      required:
        - translation
        - cached
        - cache_key
      properties:
        translation:
          type: string
          description: The translated text
          example: "γεια σας"
        cached:
          type: boolean
          description: Whether this result was served from cache (true) or freshly generated (false)
          example: false
        cache_key:
          type: string
          description: Unique cache key for this translation (hash of content + languages)
          example: "a1b2c3d4e5f67890abcdef1234567890abcdef1234567890abcdef1234567890"
      example:
        translation: "γεια σας"
        cached: false
        cache_key: "a1b2c3d4e5f67890abcdef1234567890abcdef1234567890abcdef1234567890"

    UpdateFlashcardRequest:
      type: object
      properties:
        question:
          type: string
          description: Updated question text (optional)
          example: "goodbye"
        answer:
          type: string
          description: Updated answer text (optional)
          example: "αντίο"
      example:
        question: "goodbye"
        answer: "αντίο"

    RandomFlashcardResponse:
      type: object
      required:
        - flashcard
      properties:
        flashcard:
          $ref: '#/components/schemas/Flashcard'
        ai_hint:
          type: string
          nullable: true
          description: AI-generated hint or context for the flashcard (may be null if AI is unavailable)
          example: "A common Greek greeting used in formal situations"

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Both question and answer cannot be empty"